Fix (rename backend api folder name)
