--- api/nltk_download.py
@@ -1,3 +1,0 @@
-import nltk
-nltk.download('stopwords')
-nltk.download('wordnet')
