+++ be-api/knowledge_base.py
@@ -1,0 +1,309 @@
+faq_data = {
+    "Hi": "Hello there!",
+  "Hello": "Hi!",
+  "Hey": "Hey!",
+  "Greetings": "Greetings!",
+  "Good morning": "Good morning!",
+  "Good afternoon": "Good afternoon!",
+  "Good evening": "Good evening!",
+  "Hello, how are you?": "I'm doing well, thank you for asking. How can I help you today?",
+  "How are you?": "I'm doing well, thank you.",
+  "How's it going?": "It's going well, thanks for asking.",
+  "What's up?": "Not much. How can I assist you?",
+  "Hey there": "Hey there!",
+  "Hi there": "Hi there!",
+  "Hello you": "Hello!",
+  "Good day": "Good day to you!",
+  "Nice to meet you": "It's nice to interact with you too.",
+  "Pleased to meet you": "Pleased to interact with you as well.",
+  "Welcome": "Welcome! How can I be of service?",
+  "Hi, bot": "Hello!",
+  "Hello AI": "Hello!",
+  "Hey assistant": "Hey!",
+  "Morning": "Morning!",
+  "Afternoon": "Afternoon!",
+  "Evening": "Evening!",
+  "Hi, how's it going?": "Hi! It's going well, thanks for asking.",
+  "Hello there, how are you doing?": "Hello there! I'm doing well, thank you.",
+  "Hey, what's up?": "Hey! Not much on my end. How can I help you?",
+  "Greetings, how are you today?": "Greetings! I'm doing well today, thank you.",
+    "asylum process": "The asylum process typically involves submitting an application to the relevant government agency, attending interviews, and potentially waiting for a decision. The specific steps and timelines can vary depending on the country.",
+    "legal rights refugees": "As a refugee, you generally have rights related to protection from refoulement (forced return), access to basic necessities, and the right to seek employment and education, although these can vary by location and legal status.",
+    "finding housing": "Finding housing can be challenging. You can explore options through refugee support organizations, local charities, and government assistance programs. It's often helpful to register with relevant housing services as soon as possible.",
+    "healthcare access": "Refugees are usually entitled to some level of healthcare. The extent and type of coverage depend on the country's laws. You should inquire with refugee support organizations or government agencies about how to access healthcare services.",
+    "help filling forms": "I can try to help you with general questions about forms. Please tell me which form you are working on and what specific question you have about it.",
+    "accessing food banks": "Food banks are organizations that provide free food to those in need. You can usually find information about local food banks through refugee support organizations or by searching online for 'food banks near me'.",
+    "learning the local language": "Learning the local language can greatly help with integration. Many communities offer language classes for newcomers, often through volunteer organizations or government programs. Check with local refugee centers for information.",
+    "support for children": "There are various support services available for refugee children, including educational assistance, psychosocial support, and programs aimed at helping them adjust to their new environment. Refugee support organizations are a good place to start.",
+    "feeling overwhelmed": "It's completely normal to feel overwhelmed when facing difficult circumstances. Remember to take things one step at a time and seek support from friends, family, or support organizations.",
+  "dealing with sadness": "Allow yourself to feel sadness. It's a natural response to loss and hardship. Talking to someone you trust or a mental health professional can help.",
+  "managing anxiety": "Anxiety is common in stressful situations. Try relaxation techniques like deep breathing, mindfulness, or gentle exercise. If anxiety is persistent, consider seeking professional help.",
+  "coping with fear": "Fear can be intense, especially when safety is a concern. Try to focus on what you can control and seek information from reliable sources. Connecting with others can also provide comfort.",
+  "loneliness as a refugee": "Feeling lonely after displacement is understandable. Try to connect with other refugees or community members through support groups or local organizations.",
+  "homesickness": "Missing home is a natural feeling. Try to keep memories alive through photos or stories, while also focusing on building a new sense of belonging in your current location.",
+  "frustration with the process": "Navigating bureaucratic processes can be frustrating. Try to gather information, seek assistance from caseworkers, and remember that you are not alone in this.",
+  "anger and displacement": "Anger is a valid emotion in the face of injustice or loss. Finding healthy ways to express anger, such as through writing or talking, can be helpful.",
+  "building resilience": "Resilience is the ability to bounce back from challenges. Focus on your strengths, seek support, and believe in your capacity to cope.",
+  "finding hope in difficult times": "Even in dark times, it's important to find small sources of hope. Focus on positive moments, connect with supportive people, and remember your inner strength.",
+  "importance of self-care": "Taking care of your physical and emotional well-being is crucial. Prioritize rest, healthy eating, and activities that bring you comfort.",
+  "connecting with my community": "Engaging with your local community can help you feel more connected. Look for local events, volunteer opportunities, or community centers.",
+  "dealing with trauma": "Experiences of trauma can have lasting effects. Seeking professional help from therapists or counselors specializing in trauma is essential.",
+  "support for survivors of violence": "If you have experienced violence, there are specialized support services available. Reach out to refugee organizations or mental health services for assistance.",
+  "understanding cultural differences": "Adapting to a new culture can be challenging. Be patient with yourself and try to learn about local customs and traditions.",
+  "dealing with discrimination": "Experiencing discrimination is hurtful. Seek support from advocacy groups and remember that you have rights and deserve to be treated with respect.",
+  "finding legal assistance for my situation": "Legal aid organizations can provide guidance on your rights and the legal processes relevant to your situation. Contact refugee support agencies for referrals.",
+  "accessing mental health support": "Mental health support is available for refugees and displaced people. Ask refugee organizations or healthcare providers about available services.",
+  "support groups for refugees": "Support groups can provide a safe space to connect with others who have similar experiences. Refugee agencies often facilitate these groups.",
+  "helping my children adjust": "Children may need extra support to adjust to new environments. Provide reassurance, maintain routines, and seek child-friendly support services if needed.",
+  "dealing with language barriers": "Language barriers can be isolating. Utilize translation services, language learning resources, and be patient as you learn the local language.",
+  "financial assistance for refugees": "Financial assistance programs may be available through government agencies or non-profit organizations. Inquire with refugee support services for information.",
+  "finding employment as a refugee": "Many countries offer pathways for refugees to find employment. Seek guidance from employment services specifically for newcomers.",
+  "educational opportunities for refugees": "Access to education is often a right for refugees. Explore local schools, vocational training programs, and scholarships.",
+  "registering my children in school": "Refugee support organizations can assist with the process of registering your children in local schools.",
+  "accessing transportation": "Understanding the local transportation system is important. Ask for information on public transport or assistance with travel if needed.",
+  "staying informed about my rights": "It's important to know your rights as a refugee. Seek information from legal aid organizations and reliable government sources.",
+  "connecting with interpreters": "Language interpretation services are often available through refugee agencies and during official appointments.",
+  "what to do in case of emergency": "Know the local emergency numbers and procedures. Refugee organizations can provide information on emergency support.",
+  "dealing with uncertainty about the future": "Uncertainty can be difficult. Focus on the present, take things day by day, and seek support to manage anxiety about the future.",
+  "maintaining hope for the future": "Holding onto hope is important for well-being. Connect with positive influences and focus on your goals for the future.",
+  "the importance of community solidarity": "Community support plays a vital role in helping refugees and displaced people rebuild their lives. Engage with local initiatives.",
+  "how to help other refugees": "If you are able, volunteering or supporting refugee organizations can be a meaningful way to help others.",
+  "understanding the role of empathy": "Empathy, the ability to understand and share the feelings of others, is crucial in supporting those who have experienced displacement.",
+  "practicing active listening": "When someone is sharing their experiences, active listening – fully focusing and responding thoughtfully – can be very supportive.",
+  "showing compassion": "Compassion, a feeling of concern for others, is essential in creating a welcoming and supportive environment.",
+  "the power of kindness": "Small acts of kindness can have a significant positive impact on someone who is struggling.",
+  "respecting individual experiences": "Remember that every refugee's journey and experiences are unique. Avoid making generalizations.",
+  "creating a welcoming environment": "Efforts to create inclusive and welcoming spaces can help refugees feel safer and more accepted.",
+  "addressing stigma and prejudice": "Challenging negative stereotypes and prejudice against refugees is important for fostering integration.",
+  "promoting understanding and awareness": "Raising awareness about the realities faced by refugees can help build empathy and support.",
+  "the role of cultural sensitivity": "Being aware of and respecting cultural differences is key to effective communication and support.",
+  "building trust with those who have been displaced": "Trust is essential and can be built through consistent support, honesty, and respect.",
+  "supporting the dignity of refugees": "It's crucial to treat all individuals with dignity and respect, regardless of their circumstances.",
+  "empowering refugees to rebuild their lives": "Support should aim to empower refugees to become self-sufficient and rebuild their lives with dignity.",
+  "the importance of patience in the integration process": "Integration takes time and patience from both the newcomers and the host community.",
+  "recognizing the strengths and resilience of refugees": "It's important to acknowledge the incredible strength and resilience that refugees demonstrate.",
+  "how trauma can affect emotions and behavior": "Trauma can have profound effects on emotional regulation and behavior. Understanding this is key to providing appropriate support.",
+  "the importance of creating safe spaces": "Safe and supportive environments are crucial for healing and well-being.",
+  "understanding the grieving process after loss": "Displacement often involves significant loss, and understanding the grieving process can help in providing support.",
+  "supporting mental well-being in crisis situations": "Mental health support is a critical aspect of humanitarian aid.",
+  "how to communicate effectively with someone who has experienced trauma": "Communication should be gentle, patient, and focused on creating a sense of safety.",
+  "the role of hope in recovery": "Hope for a better future is a powerful motivator in the recovery process.",
+  "connecting with resources for emotional support": "Knowing where to find mental health resources is vital.",
+  "practicing empathy in conversations": "Try to imagine yourself in the other person's situation to better understand their feelings.",
+  "validating someone's feelings": "Acknowledge and validate the emotions someone is expressing, even if you don't fully understand them.",
+  "offering practical help": "Sometimes, practical assistance with daily tasks can be more helpful than words alone.",
+  "being a reliable source of support": "Consistency and reliability in your support can build trust.",
+  "knowing your own limits as a supporter": "It's important to take care of your own well-being while supporting others.",
+  "encouraging professional help when needed": "Recognize when someone's needs exceed your capacity and encourage them to seek professional help.",
+  "the impact of displacement on identity": "Displacement can significantly impact a person's sense of identity. Supportive conversations can help.",
+  "supporting cultural identity in a new environment": "Helping refugees maintain connections to their culture can foster a sense of belonging.",
+  "navigating the legal system as a refugee in Nigeria": "Understanding the specific legal processes in Nigeria is crucial. Seek advice from legal aid organizations.",
+  "accessing education and training in Nigeria": "Explore the educational and vocational training opportunities available to refugees in Nigeria.",
+  "healthcare services available to refugees in Nigeria": "Understand the healthcare options and how to access them within Nigeria.",
+  "finding safe accommodation in Nigeria": "Seek information on safe housing options through refugee agencies and local resources.",
+  "support for unaccompanied minors in Nigeria": "Specialized support is available for children who have arrived without parents or guardians.",
+  "connecting with other refugees in Nigeria": "Support networks within the refugee community can be invaluable.",
+  "understanding the process of resettlement from Nigeria": "If resettlement to another country is a possibility, understand the procedures involved.",
+  "rights and responsibilities of refugees in Nigeria": "Be aware of your legal rights and obligations under Nigerian law.",
+  "organizations providing aid to refugees in Nigeria": "Identify and connect with local and international organizations offering assistance.",
+  "cultural norms and customs in Nigeria": "Learning about Nigerian culture can aid in smoother integration.",
+  "dealing with the climate and environment in Nigeria": "Adjusting to a new climate can be a challenge. Seek advice on staying healthy.",
+  "accessing communication and internet services in Nigeria": "Staying connected is important. Explore available communication options.",
+  "financial services available to refugees in Nigeria": "Understand if there are any specific financial assistance or banking options.",
+  "transportation options for refugees in Nigeria": "Learn about how to navigate transportation within Nigeria.",
+  "support for starting a small business in Nigeria": "If you have skills and interest, explore opportunities for entrepreneurship.",
+  "accessing translation services in Nigeria": "Know where to find reliable translation assistance.",
+  "understanding the local job market in Nigeria": "Research potential employment sectors and seek job search assistance.",
+  "support for families and children in Nigeria": "Identify resources specifically for refugee families and children.",
+  "religious and spiritual support available in Nigeria": "Connect with faith-based organizations for spiritual guidance.",
+  "dealing with potential exploitation or trafficking": "Be aware of the risks and know where to seek help if you encounter such situations.",
+  "the importance of documentation in Nigeria": "Understand the types of documents you need and how to obtain or maintain them.",
+  "what to do if you are detained or face legal issues in Nigeria": "Know your rights and where to seek legal help in case of detention.",
+  "support for LGBTQ+ refugees in Nigeria": "Understand the specific challenges and support networks available.",
+  "accessing education for adults in Nigeria": "Explore adult education and skills training programs.",
+  "dealing with food insecurity in Nigeria": "Seek assistance from food banks and other aid organizations.",
+  "understanding the role of UNHCR in Nigeria": "Learn about the mandate and services of the UN Refugee Agency.",
+  "how to report a grievance or seek help from authorities in Nigeria": "Know the proper channels for reporting issues or seeking assistance.",
+  "support for people with disabilities in Nigeria": "Identify organizations that provide support for refugees with disabilities.",
+  "accessing clean water and sanitation in Nigeria": "Understand how to access these essential services.",
+  "dealing with potential language barriers in Nigeria": "Utilize available translation and language learning resources.",
+  "understanding the local currency and economy in Nigeria": "Familiarize yourself with the Nigerian Naira and the local economic situation.",
+  "connecting with cultural organizations in Nigeria": "Engaging with cultural groups can aid in integration.",
+  "support for older refugees in Nigeria": "Identify resources that address the specific needs of older refugees.",
+  "accessing information about healthcare during emergencies in Nigeria": "Know what to do and where to go for medical help in urgent situations.",
+  "understanding the role of local communities in supporting refugees in Nigeria": "Engage with and learn from the local population.",
+  "how to stay safe and secure in your new environment in Nigeria": "Take precautions to ensure your personal safety and security.",
+  "accessing psychosocial support services in Nigeria": "Seek mental health and emotional support from available resources.",
+  "understanding the process of voluntary repatriation from Nigeria": "If returning home is an option, understand the procedures.",
+  "how to make your voice heard and advocate for your rights in Nigeria": "Learn about ways to advocate for yourself and your community.",
+  "connecting with international aid organizations in Nigeria": "Understand the roles and services of various international NGOs.",
+  "support for survivors of gender-based violence in Nigeria": "Specialized services are available for individuals who have experienced gender-based violence.",
+  "accessing childcare services in Nigeria": "If you have young children, explore available childcare options.",
+  "understanding the local customs and etiquette in Nigeria": "Respecting local customs can facilitate smoother interactions.",
+  "how to build positive relationships with the host community in Nigeria": "Engage respectfully and seek opportunities for connection.",
+  "accessing legal aid for immigration matters in Nigeria": "Seek advice from lawyers specializing in refugee and immigration law.",
+  "support for educational advancement in Nigeria": "Explore opportunities for further education and skills development.",
+  "dealing with the challenges of family reunification in Nigeria": "Understand the processes and potential difficulties of reuniting with family.",
+  "how to access information about your asylum case in Nigeria": "Know how to inquire about the status of your application.",
+  "connecting with diaspora communities in Nigeria": "Engaging with people from your home country can provide support and connection.",
+  "feeling overwhelmed": "It's completely normal to feel overwhelmed when facing difficult circumstances. Remember to take things one step at a time and seek support from friends, family, or support organizations.",
+  "dealing with sadness": "Allow yourself to feel sadness. It's a natural response to loss and hardship. Talking to someone you trust or a mental health professional can help.",
+  "managing anxiety": "Anxiety is common in stressful situations. Try relaxation techniques like deep breathing, mindfulness, or gentle exercise. If anxiety is persistent, consider seeking professional help.",
+  "coping with fear": "Fear can be intense, especially when safety is a concern. Try to focus on what you can control and seek information from reliable sources. Connecting with others can also provide comfort.",
+  "loneliness as a refugee": "Feeling lonely after displacement is understandable. Try to connect with other refugees or community members through support groups or local organizations.",
+  "homesickness": "Missing home is a natural feeling. Try to keep memories alive through photos or stories, while also focusing on building a new sense of belonging in your current location.",
+  "frustration with the process": "Navigating bureaucratic processes can be frustrating. Try to gather information, seek assistance from caseworkers, and remember that you are not alone in this.",
+  "anger and displacement": "Anger is a valid emotion in the face of injustice or loss. Finding healthy ways to express anger, such as through writing or talking, can be helpful.",
+  "building resilience": "Resilience is the ability to bounce back from challenges. Focus on your strengths, seek support, and believe in your capacity to cope.",
+  "finding hope in difficult times": "Even in dark times, it's important to find small sources of hope. Focus on positive moments, connect with supportive people, and remember your inner strength.",
+  "importance of self-care": "Taking care of your physical and emotional well-being is crucial. Prioritize rest, healthy eating, and activities that bring you comfort.",
+  "connecting with my community": "Engaging with your local community can help you feel more connected. Look for local events, volunteer opportunities, or community centers.",
+  "dealing with trauma": "Experiences of trauma can have lasting effects. Seeking professional help from therapists or counselors specializing in trauma is essential.",
+  "support for survivors of violence": "If you have experienced violence, there are specialized support services available. Reach out to refugee organizations or mental health services for assistance.",
+  "understanding cultural differences": "Adapting to a new culture can be challenging. Be patient with yourself and try to learn about local customs and traditions.",
+  "dealing with discrimination": "Experiencing discrimination is hurtful. Seek support from advocacy groups and remember that you have rights and deserve to be treated with respect.",
+  "finding legal assistance for my situation": "Legal aid organizations can provide guidance on your rights and the legal processes relevant to your situation. Contact refugee support agencies for referrals.",
+  "accessing mental health support": "Mental health support is available for refugees and displaced people. Ask refugee organizations or healthcare providers about available services.",
+  "support groups for refugees": "Support groups can provide a safe space to connect with others who have similar experiences. Refugee agencies often facilitate these groups.",
+  "helping my children adjust": "Children may need extra support to adjust to new environments. Provide reassurance, maintain routines, and seek child-friendly support services if needed.",
+  "dealing with language barriers": "Language barriers can be isolating. Utilize translation services, language learning resources, and be patient as you learn the local language.",
+  "financial assistance for refugees": "Financial assistance programs may be available through government agencies or non-profit organizations. Inquire with refugee support services for information.",
+  "finding employment as a refugee": "Many countries offer pathways for refugees to find employment. Seek guidance from employment services specifically for newcomers.",
+  "educational opportunities for refugees": "Access to education is often a right for refugees. Explore local schools, vocational training programs, and scholarships.",
+  "registering my children in school": "Refugee support organizations can assist with the process of registering your children in local schools.",
+  "accessing transportation": "Understanding the local transportation system is important. Ask for information on public transport or assistance with travel if needed.",
+  "staying informed about my rights": "It's important to know your rights as a refugee. Seek information from legal aid organizations and reliable government sources.",
+  "connecting with interpreters": "Language interpretation services are often available through refugee agencies and during official appointments.",
+  "what to do in case of emergency": "Know the local emergency numbers and procedures. Refugee organizations can provide information on emergency support.",
+  "dealing with uncertainty about the future": "Uncertainty can be difficult. Focus on the present, take things day by day, and seek support to manage anxiety about the future.",
+  "maintaining hope for the future": "Holding onto hope is important for well-being. Connect with positive influences and focus on your goals for the future.",
+  "the importance of community solidarity": "Community support plays a vital role in helping refugees and displaced people rebuild their lives. Engage with local initiatives.",
+  "how to help other refugees": "If you are able, volunteering or supporting refugee organizations can be a meaningful way to help others.",
+  "understanding the role of empathy": "Empathy, the ability to understand and share the feelings of others, is crucial in supporting those who have experienced displacement.",
+  "practicing active listening": "When someone is sharing their experiences, active listening – fully focusing and responding thoughtfully – can be very supportive.",
+  "showing compassion": "Compassion, a feeling of concern for others, is essential in creating a welcoming and supportive environment.",
+  "the power of kindness": "Small acts of kindness can have a significant positive impact on someone who is struggling.",
+  "respecting individual experiences": "Remember that every refugee's journey and experiences are unique. Avoid making generalizations.",
+  "creating a welcoming environment": "Efforts to create inclusive and welcoming spaces can help refugees feel safer and more accepted.",
+  "addressing stigma and prejudice": "Challenging negative stereotypes and prejudice against refugees is important for fostering integration.",
+  "promoting understanding and awareness": "Raising awareness about the realities faced by refugees can help build empathy and support.",
+  "the role of cultural sensitivity": "Being aware of and respecting cultural differences is key to effective communication and support.",
+  "building trust with those who have been displaced": "Trust is essential and can be built through consistent support, honesty, and respect.",
+  "supporting the dignity of refugees": "It's crucial to treat all individuals with dignity and respect, regardless of their circumstances.",
+  "empowering refugees to rebuild their lives": "Support should aim to empower refugees to become self-sufficient and rebuild their lives with dignity.",
+  "the importance of patience in the integration process": "Integration takes time and patience from both the newcomers and the host community.",
+  "recognizing the strengths and resilience of refugees": "It's important to acknowledge the incredible strength and resilience that refugees demonstrate.",
+  "how trauma can affect emotions and behavior": "Trauma can have profound effects on emotional regulation and behavior. Understanding this is key to providing appropriate support.",
+  "the importance of creating safe spaces": "Safe and supportive environments are crucial for healing and well-being.",
+  "understanding the grieving process after loss": "Displacement often involves significant loss, and understanding the grieving process can help in providing support.",
+  "supporting mental well-being in crisis situations": "Mental health support is a critical aspect of humanitarian aid.",
+  "how to communicate effectively with someone who has experienced trauma": "Communication should be gentle, patient, and focused on creating a sense of safety.",
+  "the role of hope in recovery": "Hope for a better future is a powerful motivator in the recovery process.",
+  "connecting with resources for emotional support": "Knowing where to find mental health resources is vital.",
+  "practicing empathy in conversations": "Try to imagine yourself in the other person's situation to better understand their feelings.",
+  "validating someone's feelings": "Acknowledge and validate the emotions someone is expressing, even if you don't fully understand them.",
+  "offering practical help": "Sometimes, practical assistance with daily tasks can be more helpful than words alone.",
+  "being a reliable source of support": "Consistency and reliability in your support can build trust.",
+  "knowing your own limits as a supporter": "It's important to take care of your own well-being while supporting others.",
+  "encouraging professional help when needed": "Recognize when someone's needs exceed your capacity and encourage them to seek professional help.",
+  "the impact of displacement on identity": "Displacement can significantly impact a person's sense of identity. Supportive conversations can help.",
+  "supporting cultural identity in a new environment": "Helping refugees maintain connections to their culture can foster a sense of belonging.",
+  "navigating the legal system as a refugee in Nigeria": "Understanding the specific legal processes in Nigeria is crucial. Seek advice from legal aid organizations.",
+  "accessing education and training in Nigeria": "Explore the educational and vocational training opportunities available to refugees in Nigeria.",
+  "healthcare services available to refugees in Nigeria": "Understand the healthcare options and how to access them within Nigeria.",
+  "finding safe accommodation in Nigeria": "Seek information on safe housing options through refugee agencies and local resources.",
+  "support for unaccompanied minors in Nigeria": "Specialized support is available for children who have arrived without parents or guardians.",
+  "connecting with other refugees in Nigeria": "Support networks within the refugee community can be invaluable.",
+  "understanding the process of resettlement from Nigeria": "If resettlement to another country is a possibility, understand the procedures involved.",
+  "rights and responsibilities of refugees in Nigeria": "Be aware of your legal rights and obligations under Nigerian law.",
+  "organizations providing aid to refugees in Nigeria": "Identify and connect with local and international organizations offering assistance.",
+  "cultural norms and customs in Nigeria": "Learning about Nigerian culture can aid in smoother integration.",
+  "dealing with the climate and environment in Nigeria": "Adjusting to a new climate can be a challenge. Seek advice on staying healthy.",
+  "accessing communication and internet services in Nigeria": "Staying connected is important. Explore available communication options.",
+  "financial services available to refugees in Nigeria": "Understand if there are any specific financial assistance or banking options.",
+  "transportation options for refugees in Nigeria": "Learn about how to navigate transportation within Nigeria.",
+  "support for starting a small business in Nigeria": "If you have skills and interest, explore opportunities for entrepreneurship.",
+  "accessing translation services in Nigeria": "Know where to find reliable translation assistance.",
+  "understanding the local job market in Nigeria": "Research potential employment sectors and seek job search assistance.",
+  "support for families and children in Nigeria": "Identify resources specifically for refugee families and children.",
+  "religious and spiritual support available in Nigeria": "Connect with faith-based organizations for spiritual guidance.",
+  "dealing with potential exploitation or trafficking": "Be aware of the risks and know where to seek help if you encounter such situations.",
+  "the importance of documentation in Nigeria": "Understand the types of documents you need and how to obtain or maintain them.",
+  "what to do if you are detained or face legal issues in Nigeria": "Know your rights and where to seek legal help in case of detention.",
+  "support for LGBTQ+ refugees in Nigeria": "Understand the specific challenges and support networks available.",
+  "accessing education for adults in Nigeria": "Explore adult education and skills training programs.",
+  "dealing with food insecurity in Nigeria": "Seek assistance from food banks and other aid organizations.",
+  "understanding the role of UNHCR in Nigeria": "Learn about the mandate and services of the UN Refugee Agency.",
+  "how to report a grievance or seek help from authorities in Nigeria": "Know the proper channels for reporting issues or seeking assistance.",
+  "support for people with disabilities in Nigeria": "Identify organizations that provide support for refugees with disabilities.",
+  "accessing clean water and sanitation in Nigeria": "Understand how to access these essential services.",
+  "dealing with potential language barriers in Nigeria": "Utilize available translation and language learning resources.",
+  "understanding the local currency and economy in Nigeria": "Familiarize yourself with the Nigerian Naira and the local economic situation.",
+  "connecting with cultural organizations in Nigeria": "Engaging with cultural groups can aid in integration.",
+  "support for older refugees in Nigeria": "Identify resources that address the specific needs of older refugees.",
+  "accessing information about healthcare during emergencies in Nigeria": "Know what to do and where to go for medical help in urgent situations.",
+  "understanding the role of local communities in supporting refugees in Nigeria": "Engage with and learn from the local population.",
+  "how to stay safe and secure in your new environment in Nigeria": "Take precautions to ensure your personal safety and security.",
+  "accessing psychosocial support services in Nigeria": "Seek mental health and emotional support from available resources.",
+  "understanding the process of voluntary repatriation from Nigeria": "If returning home is an option, understand the procedures.",
+  "how to make your voice heard and advocate for your rights in Nigeria": "Learn about ways to advocate for yourself and your community.",
+  "connecting with international aid organizations in Nigeria": "Understand the roles and services of various international NGOs.",
+  "support for survivors of gender-based violence in Nigeria": "Specialized services are available for individuals who have experienced gender-based violence.",
+  "accessing childcare services in Nigeria": "If you have young children, explore available childcare options.",
+  "understanding the local customs and etiquette in Nigeria": "Respecting local customs can facilitate smoother interactions.",
+  "how to build positive relationships with the host community in Nigeria": "Engage respectfully and seek opportunities for connection.",
+  "accessing legal aid for immigration matters in Nigeria": "Seek advice from lawyers specializing in refugee and immigration law.",
+  "support for educational advancement in Nigeria": "Explore opportunities for further education and skills development.",
+  "dealing with the challenges of family reunification in Nigeria": "Understand the processes and potential difficulties of reuniting with family.",
+  "how to access information about your asylum case in Nigeria": "Know how to inquire about the status of your application.",
+  "connecting with diaspora communities in Nigeria": "Engaging with people from your home country can provide support and connection.",
+  "I need immediate shelter, where can I go?": "Please contact [Name of relevant Nigerian agency/NGO for shelter] at [Phone number/website] or visit [Address of shelter information point].",
+  "Is there a safe place for me and my children tonight?": "Yes, there are emergency shelters available. Please reach out to [Name of relevant child protection agency/shelter] at [Phone number/website].",
+  "I am injured and need medical attention urgently.": "Go to the nearest hospital emergency room immediately. If possible, contact [Name of relevant medical aid organization] at [Phone number].",
+  "Where is the nearest clinic or hospital that helps refugees?": "You can find a list of healthcare facilities that assist refugees at [Website/resource link] or contact [Helpline number].",
+  "We have run out of food, where can we get emergency food supplies?": "Please contact [Name of food bank/aid organization] at [Phone number/address] for emergency food assistance.",
+  "Is there any assistance for clean drinking water?": "[Name of relevant organization] provides clean water solutions. Contact them at [Phone number/website] or visit their distribution point at [Location].",
+  "I have lost contact with my family, how can I find them?": "The [Name of tracing agency, e.g., Red Cross] may be able to help. Contact them at [Phone number/website/address].",
+  "My documents were lost, what do I need to do?": "Report the loss to the local police and then contact the [Name of relevant government agency for refugee documentation] for guidance.",
+  "I am facing legal issues, where can I get legal aid?": "[Name of legal aid organization for refugees] offers free legal assistance. Contact them at [Phone number/website].",
+  "I need help registering for assistance, who can guide me?": "Visit the registration center at [Address] or contact [Name of registration agency] at [Phone number] for assistance.",
+  "Is there any financial aid available for urgent needs?": "[Name of financial aid organization] may provide emergency financial assistance. Please inquire about eligibility.",
+  "I am a single mother with children, is there specific support for us?": "[Name of organization supporting single mothers] offers specialized programs. Contact them at [Phone number/website].",
+  "We are a large family, are there resources for larger households?": "Please inquire with [Name of aid organization] about assistance for large families.",
+  "I have a disability and need specific assistance.": "[Name of disability support organization] can provide support tailored to your needs. Contact them at [Phone number/website].",
+  "Is there any help for pregnant women or new mothers?": "[Name of maternal health organization] offers support for pregnant women and new mothers. Contact them at [Phone number/clinic address].",
+  "My child is sick, where can I get pediatric care?": "The [Name of children's clinic/hospital] has pediatric services. Visit them at [Address].",
+  "We are in a dangerous situation, who can help us evacuate?": "Contact the [Name of emergency response agency] immediately at [Emergency phone number].",
+  "I have witnessed violence and need protection.": "Please report this to the [Name of protection agency] at [Phone number/location].",
+  "Where can I report a crime or seek security assistance?": "Contact the local police at [Police emergency number] or visit the nearest police station.",
+  "I need help communicating, are there interpreters available?": "[Name of interpretation service] can provide interpreters. Contact them at [Phone number].",
+  "Is there any aid for clothing or hygiene supplies?": "[Name of aid organization providing supplies] distributes clothing and hygiene items at [Location/schedule].",
+  "We need blankets and protection from the elements.": "Please contact [Name of organization providing shelter items] for assistance.",
+  "Are there any services for people with specific medical conditions?": "Please provide more details about the condition so I can direct you to the appropriate specialized service.",
+  "I need information about my rights as a displaced person in Nigeria.": "[Name of legal aid organization] can provide information about your rights.",
+  "Where can I find information about the local area and resources?": "Visit the community center at [Address]"
+}
+
+def get_faq_answer(query):
+    query = query.lower()
+    for key, answer in faq_data.items():
+        keywords = key.lower().split()
+        if any(keyword in query for keyword in keywords):
+            return answer
+    return "I'm sorry, I don't have specific information on that topic right now. Could you rephrase your question?"
+
+def get_form_help(query):
+    query = query.lower()
+    if "form" in query:
+        if "registration" in query:
+            if "name" in query or "full name" in query:
+                return "The 'Full Name' field requires you to enter your complete legal name as it appears on your identification documents."
+            elif "date of birth" in query or "dob" in query:
+                return "The 'Date of Birth' field asks for your birth date in the format DD/MM/YYYY."
+            elif "nationality" in query:
+                return "The 'Nationality' field requires you to specify your country of citizenship."
+            else:
+                return "Okay, you're working on the Refugee Registration Form. What specific field or question are you having trouble with?"
+        else:
+            return "Please tell me the name of the form you need help with."
+    return None
