+++ be-api/nltk_data/corpora/wordnet/adj.exc
