+++ be-api/nltk_data/corpora/wordnet/adv.exc
