+++ be-api/nltk_data/corpora/wordnet/data.adv
