+++ be-api/nltk_data/corpora/wordnet/data.verb
