+++ be-api/nltk_data/corpora/wordnet/index.adv
