+++ be-api/nltk_data/corpora/wordnet/lexnames
