+++ be-api/nltk_data/corpora/wordnet/noun.exc
