+++ be-api/nltk_data/corpora/wordnet/verb.exc
