+++ be-api/nltk_download.py
@@ -1,0 +1,3 @@
+import nltk
+nltk.download('stopwords')
+nltk.download('wordnet')
